add login

+++ src/Api.Service/Services/UserService.cs
@@ -4,41 +4,75 @@ using System.Threading.Tasks;
 using Api.Domain.Entities;
 using Api.Domain.Interfaces;
 using Api.Domain.Interfaces.Services.User;
+using Api.Utils.Functions;
+using AutoMapper;
+using Domain.Dtos.User;
+using Domain.Models;
 
 namespace Api.Service.Services
 {
     public class UserService : IUserService
     {
         private IRepository<UserEntity> _repository;
+        private readonly IMapper _mapper;
 
-        public UserService(IRepository<UserEntity> repository)
+        public UserService(IRepository<UserEntity> repository, IMapper mapper)
         {
             _repository = repository;
+            _mapper = mapper;
         }
 
         public async Task<bool> Delete(Guid id)
         {
+            var userExists = await _repository.SelectAsync(id);
+
+            if (userExists == null)
+            {
+                throw new ArgumentNullException("user not exists");
+            }
+
             return await _repository.DeleteAsync(id);
         }
 
-        public async Task<UserEntity> Get(Guid id)
+        public async Task<UserDto> Get(Guid id)
         {
-            return await _repository.SelectAsync(id);
+            UserEntity entity = await _repository.SelectAsync(id);
+            return _mapper.Map<UserDto>(entity) ?? null;
         }
 
-        public async Task<IEnumerable<UserEntity>> GetAll()
+        public async Task<IEnumerable<UserDto>> GetAll()
         {
-            return await _repository.SelectAsync();
+            IEnumerable<UserEntity> listEntity = await _repository.SelectAsync();
+            return _mapper.Map<IEnumerable<UserDto>>(listEntity);
         }
 
-        public async Task<UserEntity> Post(UserEntity user)
+        public async Task<UserDtoCreateResult> Post(UserDtoCreate user)
         {
-            return await _repository.InsertAsync(user);
+            UserModel model = _mapper.Map<UserModel>(user);
+            UserEntity entity = _mapper.Map<UserEntity>(model);
+            entity.Password = PasswordHashed.HashPassword(user.Password);
+            UserEntity result = await _repository.InsertAsync(entity);
+
+            return _mapper.Map<UserDtoCreateResult>(result);
         }
 
-        public async Task<UserEntity> Put(UserEntity user)
+        public async Task<UserDtoUpdateResult> Put(UserDtoUpdate user)
         {
-            return await _repository.UpdateAsync(user);
+            var userExists = await _repository.SelectAsync(user.Id);
+
+            if (userExists == null)
+            {
+                throw new ArgumentNullException("user not exists");
+            }
+
+            UserModel model = _mapper.Map<UserModel>(user);
+            UserEntity entity = _mapper.Map<UserEntity>(model);
+
+            entity.Password = userExists.Password;
+
+            UserEntity result = await _repository.UpdateAsync(entity);
+
+            return _mapper.Map<UserDtoUpdateResult>(result);
         }
     }
 }
